requirements.txt
# Turbo.az Scraper Requirements

# Core dependencies
aiohttp>=3.9.0
asyncio>=3.4.3
beautifulsoup4>=4.12.0
pandas>=2.1.0
openpyxl>=3.1.0

# Additional utilities
lxml>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
